pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ge-pipe"
version = "1.0.0"
description = "ELT pipeline for the OSRS Grand Exchange. No bots required."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.0",
    "dagster>=1.7",
    "dagster-webserver>=1.7",
    "dagster-dbt>=1.7",
    "dbt-postgres>=1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-httpx>=0.30",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ge_pipe*"]
